Fix bug in cheers app that prevents user cheers
Currently if a user sends a cheer that has words that contain their
username, the cheer is blocked. Also different users that have names
where one is a subset of another would also be blocked. This commit aims
to rectify that by building a list of the recipients and checking for
the username.